some codes were modified
