+++ reasoners/tools/__init__.py
@@ -1,0 +1,1 @@
+from .tools import wikisearch, wikilookup, wikifinish
